Suse manager settings moved under settings controller
