+++ lib/trento_web/views/v1/settings_view.ex
@@ -38,4 +38,18 @@ defmodule TrentoWeb.V1.SettingsView do
       }
     }
   end
+
+  def render("suma_credentials.json", %{
+        settings: %{
+          url: url,
+          username: username,
+          ca_uploaded_at: ca_uploaded_at
+        }
+      }) do
+    %{
+      url: url,
+      username: username,
+      ca_uploaded_at: ca_uploaded_at
+    }
+  end
 end
